Note that some content has been copied from elsewhere
But it'd be difficult to cite all the authors and apply e.g. a Creative
Commons license. For such a small text it shouldn't be a major problem,
the goal is to spread information about GLib/GTK+.

The authors who have written the original content on gtk.org or the API
references normally agree that the text is copied elsewhere, for the own
good of GTK+. It'd be useless to reinvent the wheel, the copied
paragraphs were copied because it's well explained and I couldn't do
better.